Major code update.
Fixed:
[UI]\>
Fixed memory leaks when switching game modes and selecting maps >Fixed a critical engine error when using a map as a background in the menu.

[Engine]\>
Increased the base engine resource limits.

[Zombie Evolution]\>
Fixed a pop-up message that appeared when using zombie evolution.

[Gun Death]\>
Fixed the level-based weapon distribution.

Added:
[Gun Death]\>
Added RespawnBar displayed after death.

[Heaven Splitter]\>
Using the CSMOE-Citrus code, a mine was added to the alternative attack.

[Twin Shadow Axes]\>
The R+MOUSEATTACK1 key combination was added to spawn an entity.